+++ srcs/calc_combinations.c
@@ -6,7 +6,7 @@
 /*   By: lravier <lravier@student.codam.nl>           +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/05/22 16:03:32 by lravier       #+#    #+#                 */
-/*   Updated: 2020/06/29 20:11:54 by kim           ########   odam.nl         */
+/*   Updated: 2020/07/07 14:07:50 by lravier       ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,11 +21,11 @@ static size_t	ft_factorial(size_t n)
 
 // N = number of paths, r is max number of parallel paths
 
-ssize_t			calc_combinations(const size_t n, size_t r)
+long long unsigned			calc_combinations(const size_t n, size_t r)
 {
-	size_t	result;
-	size_t	fact_n;
-	size_t	i;
+	long long unsigned	result;
+	long long unsigned	fact_n;
+	long long unsigned	i;
 
 	// (n!) / (r!(n-r)!);
 	result = 0;
